priority column

+++ resources/views/attribute_values/index.blade.php
@@ -114,7 +114,7 @@
         $(document).ready(function(){
             // Data table for serverside
             $('#attribute-values').DataTable({
-                "pageLength": 25,
+                "pageLength": 40,
                 "order": [[ 0, 'desc' ]],
                 "processing": true,
                 "serverSide": true,
@@ -151,10 +151,23 @@
 
      $('.checkboxes').removeAttr('checked');    
  });
- $('#deleteTrigger').on('click',function () {    
-     $('#delete_all').submit();    
- });
- 
+ $('#deleteTrigger').on('click',function () { 
+    swal({
+        title: 'Are you sure?',
+        text: 'You won\'t be able to revert this!',
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!',
+        closeOnConfirm: true,
+        closeOnCancel: true
+    }).then(function (isConfirm) {
+        if (isConfirm.value) {
+          $('#delete_all').submit();        
+        }
+    }).catch(swal.noop)
+  });
 
         </script>
 @endsection
